Quit to long convert

+++ composeApp/src/commonMain/kotlin/com/redfrogec/credigo/ui/viewModel/ActiveLoansViewModel.kt
@@ -42,7 +42,7 @@ class ActiveLoansViewModel(private val loanSDK: LoanSDK, private val chargeSDK: 
 
     fun loadLoans() {
         val userId = settings.getLong(Constants.USER_ID, 0)
-        val activeLoans = loanSDK.selectAllLoansByUserId(userId.toLong(), true)
+        val activeLoans = loanSDK.selectAllLoansByUserId(userId, true)
         if (activeLoans.isNotEmpty()) {
             activeLoans.forEach { loan ->
                 val chargesPaid = chargeSDK.selectChargePaid(loan.id)
